chore: rename ag grid as its not lazy loaded
