+++ libs/datagrid/src/lib/ag-grid/ag-grid.tsx
@@ -1,12 +1,12 @@
 import { forwardRef } from 'react';
 import type { AgGridReactProps, AgGridReact } from 'ag-grid-react';
-import { AgGridThemed } from './ag-grid-lazy-themed';
+import { AgGridThemed } from './ag-grid-themed';
 
 type Props = AgGridReactProps & {
   style?: React.CSSProperties;
   gridRef?: React.Ref<AgGridReact>;
 };
 
-export const AgGridLazy = forwardRef<AgGridReact, Props>((props, ref) => (
+export const AgGrid = forwardRef<AgGridReact, Props>((props, ref) => (
   <AgGridThemed {...props} gridRef={ref} />
 ));
